Move api.helpers test to separated file
